% Generated by roxygen2 (4.0.2): do not edit by hand
\name{acf2}
\alias{acf2}
\title{Function for computing autocorrelations}
\usage{
acf2(x, lag.max = 10, type = c("correlation", "covariance", "partial"),
  plot = FALSE, na.action = na.omit, demean = TRUE, ...)
}
\description{
By default, this function returns the lag-10 autocorrelations of a
numeric vector and omits missing values.
}
\examples{
x <- rnorm(100)
x[5] <- NA
acf2(x)
}

