\name{bafArray}
\alias{bafArray}
\docType{data}
\title{B Allele Frequencies}
\description{This dataset contains B Allele Frequencies across all trio samples}
\usage{data(bafArray)}
\format{
 Rows are markers.
 Columns are the samples making up the trios. 
}
\source{
HapMap Phase 2 data
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(bafArray)
str(bafArray)
}
\keyword{datasets}
