% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculateMindist}
\alias{calculateMindist}
\alias{calculateMindist,arrayORff_array-method}
\alias{calculateMindist,matrix-method}
\title{Deprecated function to calculate the minimum distance}
\usage{
calculateMindist(object, ...)
}
\arguments{
\item{object}{see \code{showMethods("calculateMindist")}}

\item{...}{Ignored}
}
\description{
Deprecated function to calculate the minimum distance
}

