% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion-methods.R
\docType{methods}
\name{coerce}
\alias{coerce}
\alias{coerce,TrioSet,TrioSetList-method}
\alias{coerce,TrioSet,data.frame-method}
\alias{stack,TrioSetList-method}
\alias{stack,TrioSetLiset-method}
\title{Coerces a TrioSetList to a TrioSet}
\usage{
\S4method{stack}{TrioSetList}(x, ...)
}
\arguments{
\item{x}{a \code{TrioSetList}}

\item{...}{ignored}
}
\value{
a \code{TrioSet}
}
\description{
Coerces a TrioSetList to a TrioSet
}
