\name{Deprecated}
\title{MinimumDistance Deprecated}
\alias{xyplotTrioLrrBaf}
\alias{computeBayesFactor}
\alias{computeBayesFactor,TrioSet-method}
\alias{computeBayesFactor,TrioSetList-method}
\alias{gcSubtract,TrioSet-method}
\alias{gcSubtract,TrioSetList-method}
\alias{narrow}
\description{ The function, class, or data object you asked for has been
  deprecated.}
\keyword{internal}