% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Pedigree-class}
\alias{Pedigree-class}
\alias{offspringNames}
\alias{offspringNames<-}
\alias{pedigree}
\alias{pedigree<-}
\alias{trios}
\title{Deprecated class for storing pedigree data}
\usage{
offspringNames(object)

offspringNames(object) <- value

trios(object)

pedigree(object)

pedigree(object) <- value
}
\arguments{
\item{object}{}

\item{value}{}
}
\description{
Deprecated class for storing pedigree data

These functions will be defunct in a future release.
}
\section{Slots}{

\describe{
\item{\code{trios}}{a \code{data.frame} with colnames 'F', 'M', and 'O'
containing sample identifiers for the father (F), mother (M), and
offspring (O).}

\item{\code{trioIndex}}{a \code{data.frame}}
}}
\examples{
trios(Pedigree())
}

