\name{fData}
\alias{fData}
\docType{data}
\title{Feature data for the trios}
\description{Feature data such as chromosome and position for each snp}
\usage{data(fData)}
\format{
 An Annotated Data Frame containing information such as chromosome and
 position for each snp.
}
\examples{
data(fData)
}
\keyword{datasets}
