% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filterExperiment}
\alias{filterExperiment}
\title{Methods for filtering MinDistExperiment objects}
\usage{
filterExperiment(object, granges, param)
}
\arguments{
\item{object}{A \code{MinDistExperiment}}

\item{granges}{A \code{GRanges}, \code{GRangesList}, or \code{MinDistGRanges} object}

\item{param}{a \code{MinDistParam} object}
}
\value{
a \code{MinDistExperiment}
}
\description{
Filter a MinDistExperiment object to exclude markers with missing
values in the low-level summaries, exclude markers that lie in
segments (\code{granges} argument) with small minimum distance
values (unlikely to be de novo)
}

