<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2: do not edit by hand
>>>>>>> upstream/master
% Please edit documentation in R/AllClasses.R, R/methods-FilterParamMD.R
\docType{class}
\name{FilterParamMD-class}
\alias{FilterParamMD-class}
\alias{show,FilterParamMD-method}
\title{A class for filtering genomic intervals called by MinimumDistance}
\usage{
\S4method{show}{FilterParamMD}(object)
}
\arguments{
\item{object}{a \code{FilterParamMD} object}
}
\description{
Options for filtering include the number of markers spanned by a
segment, the posterior probability of the maximum a posteriori
estimate of the trio copy number state, and the trio copy number
state.
}
\seealso{
denovo
}

