\name{logRArray}
\alias{logRArray}
\docType{data}
\title{logR Ratios}
\description{This dataset contains logR Ratios across all trio samples}
\usage{data(logRArray)}
\format{
 Rows are markers.
 Columns are the samples making up the trios. 
}
\source{
HapMap Phase 2 data
}
\examples{
data(logRArray)
str(logRArray)
}
\keyword{datasets}
