% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\name{md_gr}
\alias{md_gr}
\title{An example \code{MinDistGRanges} object}
\format{a \code{MinDistGRanges} object}
\description{
Prior to inferring de novo trio copy number states, the log R
ratios are segmented independently for each individual in a
\code{ParentOffsping} class. The segmentation results are recorded
in separate \code{GRanges} objects for the parents. For
segmentation of the offspring log R ratios and the minimum
distance, the segments are stored in separate \code{GRangesList}
objects.  For convenience, these \code{GRanges},
\code{GRangesList}, and pedigree information are bound in a single
container referred to as a \code{MinDistGRanges} object.  The
example \code{MinDistGRanges} object provided in this package was
obtained from the segmentation of the data stored in the example
\code{MinDistExperiment} object.
}

