<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2: do not edit by hand
>>>>>>> upstream/master
% Please edit documentation in R/AllClasses.R, R/methods-MDRanges.R
\docType{class}
\name{MDRanges-class}
\alias{MDRanges}
\alias{MDRanges-class}
<<<<<<< HEAD
\alias{reduce,MDRanges-method}
\title{A \code{GRanges}-derived class}
\usage{
MDRanges(..., posteriors)

\S4method{reduce}{MDRanges}(x, ...)
=======
\title{A \code{GRanges}-derived class}
\usage{
MDRanges(..., posteriors)
>>>>>>> upstream/master
}
\arguments{
\item{...}{additional arguments to \code{GRanges} constructor}

\item{posteriors}{a \code{DataFrame}}
<<<<<<< HEAD

\item{x}{a \code{MDRanges} object}
=======
>>>>>>> upstream/master
}
\description{
Contains maximum a posteriori estimates for each genomic interval
}
\examples{
MDRanges()
}

