\name{methodsForClassesNotExported}
\docType{methods}
\alias{[,LikSet-method}
\alias{[,ff_array-method}
\alias{baf,LogRatioSet-method}
\alias{baf<-,LogRatioSet-method}
\alias{logR,LogRatioSet-method}
\alias{logR<-,LogRatioSet-method}

\title{Methods for class that are not exported}

\description{Methods for classes that are not exported}

\keyword{methods}