% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MinDistExperiment}
\alias{MinDistExperiment}
\title{Constructor for \code{MinDistExperiment} class}
\usage{
MinDistExperiment(object = ArrayViews(), pedigree = ParentOffspring(), ...)
}
\arguments{
\item{object}{see \code{showMethods(MinDistExperiment)}}

\item{pedigree}{a \code{ParentOffspring} object}
}
\value{
an object of class \code{MinDistExperiment}
}
\description{
Constructor for \code{MinDistExperiment} class
}

