% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MinDistGRanges-class}
\alias{MinDistGRanges-class}
\title{A container for storing segmentation data for members in a
\code{ParentOffspring} family}
\description{
A container for storing segmentation data for members in a
\code{ParentOffspring} family
}
\section{Slots}{

\describe{
\item{\code{mindist}}{a \code{GRangesList} object}

\item{\code{offspring}}{a \code{GRangesList} object}

\item{\code{father}}{a \code{GRanges} object}

\item{\code{mother}}{a \code{GRanges} object}

\item{\code{pedigree}}{a \code{ParentOffspring} object}
}}
\examples{
data(md_gr)
offspring(md_gr)
father(md_gr)
mother(md_gr)
mindist(md_gr)
}

