% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ParentOffspring-class}
\alias{ParentOffspring}
\alias{ParentOffspring-class}
\alias{father}
\alias{mother}
\alias{offspring}
\title{Object containing the sample identifiers for members in a pedigree}
\usage{
offspring(object)

mother(object)

father(object)

ParentOffspring(id = character(), father = character(),
  mother = character(), offspring = character(), filePaths = character())
}
\arguments{
\item{id}{length-one character vector providing a family-level id}

\item{father}{length-one character vector providing sample ids for father}

\item{mother}{length-one character vector providing sample ids for mother}

\item{offspring}{offspring character vector providing sample ids for offspring (can have length greater than one if there is more than one offspring)}

\item{filePaths}{filePaths character vector providing path to low-level data}
}
\description{
Container for registering sample identifiers with membership in a
pedigree.  For representing multiple pedigrees, see
\code{\linkS4class{ParentOffspringList}}.

Accessor for the sample identifiers for the members in a
pedigree

Constructor for ParentOffspring class
}
\section{Slots}{

\describe{
\item{\code{id}}{length-one character vector providing a family-level id}

\item{\code{father}}{length-one character vector providing sample ids for father}

\item{\code{mother}}{length-one character vector providing sample ids for mother}

\item{\code{offspring}}{character vector providing sample ids for offspring (can have length greater than one if there is more than one offspring)}

\item{\code{filePaths}}{character vector providing path to low-level data}
}}
\examples{
ParentOffspring()
}
\seealso{
ParentOffspringList-class
}

