% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ParentOffspringList-class}
\alias{ParentOffspringList}
\alias{ParentOffspringList-class}
\title{A list of \code{ParentOffspring} objects}
\usage{
ParentOffspringList(pedigrees = list())
}
\arguments{
\item{pedigrees}{a list of \code{ParentOffspring} objects}
}
\description{
Each element of the list is an element of class
\code{\linkS4class{ParentOffspring}}.

Constructor for ParentOffspringList class
}
\section{Slots}{

\describe{
\item{\code{id}}{a character vector of identifiers for the
pedigrees. \code{id} must have the same length as \code{pedigrees}}

\item{\code{pedigrees}}{A list of ParentOffspring objects.}
}}
\examples{
ParentOffspringList()
}

