\name{pData}
\alias{pData}
\docType{data}
\title{Phenotypic data for all trios}
\description{Phenotypic data for all trios arranged as a 3 dimensionl array}
\usage{data(pData)}
\format{
An array with the following three dimensions number of trios X number
of columns in samplesheet X 3. Each row represents the phenoData for
one trio.
}
\source{
 HapMap Phase 2 Data
}
\examples{
data(pData)
str(pData)
}
\keyword{datasets}
