<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2: do not edit by hand
>>>>>>> upstream/master
% Please edit documentation in R/AllGenerics.R
\name{pedigreeName}
\alias{pedigreeName}
\title{Accessor for pedigree name}
\usage{
pedigreeName(object)
}
\arguments{
\item{object}{a \code{ParentOffspring} or \code{ParentOffspringList} object}
}
\description{
Accessor for pedigree name
}
\seealso{
\code{\linkS4class{ParentOffspring}} \code{\linkS4class{ParentOffspringList}}
}

