% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ILimit.R
\name{range,ILimit-method}
\alias{range,ILimit-method}
\title{compute the range of an ILimit instance}
\usage{
\S4method{range}{ILimit}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{a \code{ILimit} object}

\item{...}{ignored}

\item{na.rm}{logical. If TRUE, missing values are removed.}
}
\description{
The range method for class ILimit is used internally in
MinimumDistance.
}
