% Generated by roxygen2 (4.0.2): do not edit by hand
\name{segment2}
\alias{segment2}
\title{A wrapper for DNAcopy's segment function}
\usage{
segment2(object, ...)
}
\arguments{
\item{object}{see \code{showMethods{segment2}}}

\item{...}{Additional arguments passed to DNAcopy's \code{segment}.}
}
\description{
Methods for circular binary segmentation.
}
\seealso{
\code{\link{segment}}
}

