% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{TrioSetList-class}
\alias{$,TrioSetList-method}
\alias{TrioSetList-class}
\title{Deprecated class for storing low-level genomic data for trios}
\usage{
\S4method{$}{TrioSetList}(x, name)
}
\description{
This class is deprecated and will be defunct in a future release.
}
\section{Slots}{

\describe{
\item{\code{fatherPhenoData}}{\code{AnnotatedDataFrame} containing covariates for the father}

\item{\code{motherPhenoData}}{\code{AnnotatedDataFrame} containing covariates for the mother}

\item{\code{Pedigree}}{an object of class \code{Pedigree}}
}}

