<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2: do not edit by hand
>>>>>>> upstream/master
% Please edit documentation in R/defunct.R
\name{Defunct}
\alias{Defunct}
\alias{coerce,RangedDataCNV,GRanges-class}
\title{Defunct functions/classes/methods in the MinimumDistance package}
\description{
The function, class, or data object you asked is defunct.
}
\keyword{internal}

