\name{MinimumDistance-Defunct}
\alias{xyplotTrioLrrBaf}
\alias{computeBayesFactor}
\alias{computeBayesFactor,TrioSet-method}
\alias{computeBayesFactor,TrioSetList-method}
\alias{gcSubtract,TrioSet-method}
\alias{gcSubtract,TrioSetList-method}
\alias{narrow}

\title{Defunct functions in package \sQuote{MinimumDistance}}

\description{ These functions are defunct and no longer available.}

\details{
  Defunct functions are: \code{xyplotTrioLrrBaf},
  \code{computeBayesFactor},
  \code{gcSubtract}, \code{narrow}.
}
\keyword{internal}