<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2: do not edit by hand
>>>>>>> upstream/master
% Please edit documentation in R/datasets.R
\name{exampleTrioSetList}
\alias{exampleTrioSetList}
\alias{trioSetList}
\title{An example \code{TrioSetList} object}
\format{a \code{TrioSetList}}
\description{
A dataset containing log R ratios and B allele frequencies for two
chromosomes, organized as a \code{TrioSetList}.  Each element in
the list class is a \code{TrioSet} object.  Both \code{TrioSetList}
and \code{TrioSet} classes are deprecated; the example data will be
removed in a future release.
}

