% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mad-methods.R
\docType{methods}
\name{mad2}
\alias{mad2}
\alias{mad2,list-method}
\alias{mad2,TrioSetList-method}
\alias{mad2,matrix-method}
\alias{mad2,array-method}
\title{Deprecated wrapper for computing the median absolute deviation of
low-level summaries}
\usage{
mad2(object, byrow = FALSE, ...)

\S4method{mad2}{list}(object, byrow, pedigree, ...)

\S4method{mad2}{TrioSetList}(object, byrow = FALSE, ...)

\S4method{mad2}{matrix}(object, byrow, pedigree, ...)

\S4method{mad2}{array}(object, byrow, pedigree, ...)
}
\arguments{
\item{object}{see \code{showMethods("mad2")}}

\item{byrow}{logical if TRUE, compute the median absolute deviation of the rows of a matrix}

\item{...}{additional arguments to \code{\link{mad}}}

\item{pedigree}{an object of class \code{Pedigree}}
}
\description{
Deprecated wrapper for computing the median absolute deviation of
low-level summaries
}
