% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mad2}
\alias{mad2}
\title{Wrapper for computing the median absolute deviation of low-level summaries}
\usage{
mad2(object, byrow = FALSE, ...)
}
\arguments{
\item{object}{see \code{showMethods("mad2")}}

\item{byrow}{logical if TRUE, compute the median absolute deviation of the rows of a matrix}

\item{...}{additional arguments to \code{\link{mad}}}
}
\description{
Wrapper for computing the median absolute deviation of low-level summaries
}

