<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2: do not edit by hand
>>>>>>> upstream/master
% Please edit documentation in R/methods-grid.R
\name{mdLegend}
\alias{mdLegend}
\title{Text summary of information encapculated in a MDRanges object for a particular interval}
\usage{
mdLegend(g)
}
\arguments{
\item{g}{a \code{MDRanges} object}
}
\description{
Text summary of information encapculated in a MDRanges object for a particular interval
}

