% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-MDRanges.R
\docType{class}
\name{MDRanges-class}
\alias{MDRanges}
\alias{MDRanges-class}
\alias{reduce,MDRanges-method}
\title{A \code{GRanges}-derived class}
\usage{
MDRanges(..., posteriors)

\S4method{reduce}{MDRanges}(x, ...)
}
\arguments{
\item{...}{additional arguments to \code{GRanges} constructor}

\item{posteriors}{a \code{DataFrame}}

\item{x}{a \code{MDRanges} object}
}
\description{
Contains maximum a posteriori estimates for each genomic interval
}
\examples{
MDRanges()
}

