% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-MDRanges.R
\docType{class}
\name{MDRanges-class}
\alias{MDRanges-class}
\alias{MDRanges}
\title{A \code{GRanges}-derived class}
\usage{
MDRanges(..., posteriors)
}
\arguments{
\item{...}{additional arguments to \code{GRanges} constructor}

\item{posteriors}{a \code{DataFrame}}
}
\description{
Contains maximum a posteriori estimates for each genomic interval
}
\examples{
MDRanges()
}
