\name{methodsForClassesNotExported}
\docType{methods}
\alias{[,LikSet-method}
\alias{[,ff_array-method}
\alias{baf,LogRratioSet-method}
\alias{baf<-,LogRratioSet-method}
\alias{initialize,LogRratioSet-method}
\alias{lrr,LogRratioSet-method}
\alias{lrr<-,LogRratioSet-method}

\title{Methods for class that are not exported}

\description{Methods for classes that are not exported}

\keyword{methods}