% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{MinimumDistance}
\alias{MinimumDistance}
\alias{MinimumDistance-package}
\title{De novo copy number alterations in parent-offspring trios}
\description{
De novo copy number alterations in parent-offspring trios
}

