<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2: do not edit by hand
>>>>>>> upstream/master
% Please edit documentation in R/AllGenerics.R
\name{nMAD}
\alias{nMAD}
\alias{nMAD<-}
\title{Setter and getter for number of median absolute deviations the mean
minimum distance of a genomic interval is from zero}
\usage{
nMAD(object)

nMAD(object) <- value
}
\arguments{
\item{object}{see \code{showMethods("nMAD")}}

\item{value}{a length-one numeric vector}
}
\description{
Setter and getter for number of median absolute deviations the mean
minimum distance of a genomic interval is from zero
}

