% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-ParentOffspringList.R
\docType{class}
\name{ParentOffspringList-class}
\alias{ParentOffspringList-class}
\alias{ParentOffspringList}
\alias{pedigreeName,ParentOffspringList-method}
\alias{show,ParentOffspringList-method}
\alias{[[,ParentOffspringList,ANY,ANY-method}
\alias{"[[",ParentOffspringList,ANY,ANY-method}
\alias{[,ParentOffspringList,ANY,ANY,ANY-method}
\alias{[,ParentOffspringList,ANY-method}
\alias{length,ParentOffspringList-method}
\title{A list of \code{ParentOffspring} objects}
\usage{
ParentOffspringList(pedigrees = list(), id)

\S4method{pedigreeName}{ParentOffspringList}(object)

\S4method{show}{ParentOffspringList}(object)

\S4method{[[}{ParentOffspringList,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{ParentOffspringList,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{length}{ParentOffspringList}(x)
}
\arguments{
\item{pedigrees}{a list of \code{ParentOffspring} objects}

\item{id}{identifier for a pedigree}

\item{object}{a \code{ParentOffspringList} object}

\item{x}{a \code{ParentOffspringList} object}

\item{i}{a numeric vector for subsetting the list (optional)}

\item{j}{ignored}

\item{...}{ignored}

\item{drop}{ignored}
}
\description{
Each element of the list is an element of class
\code{\linkS4class{ParentOffspring}}.
}
\section{Slots}{

\describe{
\item{\code{id}}{a character vector of identifiers for the
pedigrees. \code{id} must have the same length as \code{pedigrees}}

\item{\code{pedigrees}}{A list of ParentOffspring objects.}
}}

\examples{
ParentOffspringList()
}
