\name{pedigreeExample}
\alias{pedigreeExample}
\docType{data}
\title{Pedigree data for 2 trios}
\description{Pedigree data for 2 trios}
\usage{data(pedigreeExample)}
\format{
  object of class \code{Pedigree}
}
\details{
  Container for storing pedigree information
}
\source{HapMap}
\seealso{\code{\linkS4class{Pedigree}}}
\examples{
data(pedigreeExample)
}
\keyword{datasets}
