% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-grid.R
\name{pedigreeViewports}
\alias{pedigreeViewports}
\title{Default viewports for plotting log R ratios, BAFs, chromosome
idiogram, and a legend for a case-parent trio}
\usage{
pedigreeViewports()
}
\description{
Default viewports for plotting log R ratios, BAFs, chromosome
idiogram, and a legend for a case-parent trio
}
\examples{
vps <- pedigreeViewports()
}
\seealso{
\code{\link{plotDenovo}} \code{\link{pedigreeGrid}}
}
