\name{samplesheetExample}
\alias{samplesheetExample}
\docType{data}
\title{
Samplesheet File
}
\description{
Example of a Samplesheet File. Contains phenoData such as Sample Names,
Gender, Group, etc. for the trios
}
\usage{data(samplesheetExample)}
\source{
HapMap Phase 2 Data
}
\examples{
data(samplesheetExample)
}
\keyword{datasets}
