% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TrioSetList}
\alias{TrioSetList}
\title{Constructor for \code{TrioSetList} class}
\usage{
TrioSetList(chromosome = integer(), pedigreeData = Pedigree(), sample.sheet,
  row.names = NULL, lrr, baf, featureData, cdfname, ffname = "", genome)
}
\description{
The \code{TrioSetList} class has been deprecated and may be removed in
a future release. Use \code{MinDistExperiment} instead.
}

